修复bug:获取indexField字段信息时boost属性类型转换异常

+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/client/BuildTool.java
@@ -2390,7 +2390,7 @@ public abstract class BuildTool {
 		indexField.setIgnoreAbove(ResultUtil.intValue(fieldInfo.get("ignore_above"),null));
 		indexField.setAnalyzer((String)fieldInfo.get("analyzer"));
 		indexField.setNormalizer((String)fieldInfo.get("normalizer"));
-		indexField.setBoost((Integer)fieldInfo.get("boost"));
+		indexField.setBoost(fieldInfo.get("boost"));
 		indexField.setCoerce(parseBoolean( fieldInfo.get("coerce")));
 		indexField.setCopyTo((String)fieldInfo.get("copy_to"));
 		indexField.setDocValues(parseBoolean(fieldInfo.get("doc_values")));//setCoerce();
